fix: remove range segments mixin, include in range
