+++ src/js/media-chrome-range.js
@@ -1,6 +1,7 @@
 import { MediaStateReceiverAttributes } from './constants.js';
 import { globalThis, document } from './utils/server-safe-globals.js';
-import { getOrInsertCSSRule, getPointProgressOnLine, cachedBoundingClientRect } from './utils/element-utils.js';
+import { insertCSSRule, getOrInsertCSSRule, getPointProgressOnLine, cachedBoundingClientRect } from './utils/element-utils.js';
+import { observeResize, unobserveResize } from './utils/resize-observer.js';
 
 const template = document.createElement('template');
 template.innerHTML = /*html*/`
@@ -189,6 +190,38 @@ template.innerHTML = /*html*/`
     :host([disabled]) #thumb {
       background-color: #777;
     }
+
+    .segments #appearance {
+      height: var(--media-range-segment-hover-height, 7px);
+    }
+
+    #background,
+    #track {
+      clip-path: url(#segments-clipping);
+    }
+
+    #segments {
+      --segments-gap: var(--media-range-segments-gap, 2px);
+      position: absolute;
+      width: 100%;
+      height: 100%;
+    }
+
+    #segments-clipping {
+      transform: translateX(calc(var(--segments-gap) / 2));
+    }
+
+    #segments-clipping:empty {
+      display: none;
+    }
+
+    #segments-clipping rect {
+      height: var(--media-range-track-height, 4px);
+      y: calc((var(--media-range-segment-hover-height, 7px) - var(--media-range-track-height, 4px)) / 2);
+      transition: var(--media-range-segment-transition, transform .1s ease-in-out);
+      transform: var(--media-range-segment-transform, scaleY(1));
+      transform-origin: center;
+    }
   </style>
   <div id="leftgap"></div>
   <div id="container">
@@ -202,6 +235,7 @@ template.innerHTML = /*html*/`
         <div id="progress"></div>
       </div>
       <div id="thumb"></div>
+      <svg id="segments"><clipPath id="segments-clipping"></clipPath></svg>
     </div>
     <input id="range" type="range" min="0" max="1" step="any" value="0">
   </div>
@@ -267,6 +301,7 @@ class MediaChromeRange extends globalThis.HTMLElement {
   #startpoint;
   #endpoint;
   #cssRules = {};
+  #segments = [];
 
   static get observedAttributes() {
     return [
@@ -355,6 +390,7 @@ class MediaChromeRange extends globalThis.HTMLElement {
     this.shadowRoot.addEventListener('focusout', this.#onFocusOut);
 
     this.#enableUserEvents();
+    observeResize(this.container, this.#updateComputedStyles);
   }
 
   disconnectedCallback() {
@@ -366,6 +402,13 @@ class MediaChromeRange extends globalThis.HTMLElement {
 
     this.shadowRoot.removeEventListener('focusin', this.#onFocusIn);
     this.shadowRoot.removeEventListener('focusout', this.#onFocusOut);
+    unobserveResize(this.container, this.#updateComputedStyles);
+  }
+
+  #updateComputedStyles = () => {
+    // This fixes a Chrome bug where it doesn't refresh the clip-path on content resize.
+    const clipping = this.shadowRoot.querySelector('#segments-clipping');
+    if (clipping) clipping.parentNode.append(clipping);
   }
 
   updatePointerBar(evt) {
@@ -376,6 +419,60 @@ class MediaChromeRange extends globalThis.HTMLElement {
     const rangePercent = this.range.valueAsNumber * 100;
     this.#cssRules.progress.style.setProperty('width', `${rangePercent}%`);
     this.#cssRules.thumb.style.setProperty('left', `${rangePercent}%`);
+  }
+
+  updateSegments(segments) {
+    const clipping = this.shadowRoot.querySelector('#segments-clipping');
+    clipping.textContent = '';
+
+    if (!segments?.length) return;
+
+    this.#cssRules.activeSegment = insertCSSRule(this.shadowRoot, '#segments-clipping rect:nth-child(0)');
+
+    this.container.classList.toggle('segments', !!segments.length);
+
+    const normalized = [...new Set([
+      +this.range.min,
+      ...segments.flatMap(s => [s.start, s.end]),
+      +this.range.max
+    ])];
+
+    this.#segments = [...normalized];
+
+    const lastMarker = normalized.pop();
+    for (const [i, marker] of normalized.entries()) {
+      const [isFirst, isLast] = [i === 0, i === normalized.length - 1];
+      const x = isFirst ? 'calc(var(--segments-gap) / -1)' : `${marker * 100}%`;
+      const x2 = isLast ? lastMarker : normalized[i + 1];
+      const width = `calc(${(x2 - marker) * 100}%${isFirst || isLast ? '' : ` - var(--segments-gap)`})`;
+
+      const segmentEl = document.createElementNS('http://www.w3.org/2000/svg', 'rect');
+      const cssRule = getOrInsertCSSRule(this.shadowRoot, `#segments-clipping rect:nth-child(${i + 1})`);
+      cssRule.style.setProperty('x', x);
+      cssRule.style.setProperty('width', width);
+      clipping.append(segmentEl);
+    }
+  }
+
+  #updateActiveSegment(evt) {
+    const rule = this.#cssRules.activeSegment;
+    if (!rule) return;
+
+    const pointerRatio = this.getPointerRatio(evt);
+    const segmentIndex = this.#segments.findIndex((start, i, arr) => {
+      const end = arr[i + 1];
+      return end != null && pointerRatio >= start && pointerRatio <= end;
+    });
+
+    const selectorText = `#segments-clipping rect:nth-child(${segmentIndex + 1})`;
+
+    if (rule.selectorText != selectorText || !rule.style.transform) {
+      rule.selectorText = selectorText;
+      rule.style.setProperty(
+        'transform',
+        'var(--media-range-segment-hover-transform, scaleY(2))'
+      );
+    }
   }
 
   getPointerRatio(evt) {
@@ -457,11 +554,13 @@ class MediaChromeRange extends globalThis.HTMLElement {
     globalThis.window?.removeEventListener('pointermove', this);
     this.toggleAttribute('dragging', false);
     this.range.disabled = this.hasAttribute('disabled');
+    this.#cssRules.activeSegment?.style.removeProperty('transform');
   }
 
   #handlePointerMove(evt) {
     this.toggleAttribute('dragging', evt.buttons === 1 || evt.pointerType !== 'mouse');
     this.updatePointerBar(evt);
+    this.#updateActiveSegment(evt);
 
     // If the native input target & events are used don't fire manual input events.
     if (this.dragging && (evt.pointerType !== 'mouse' || !this.#isInputTarget)) {
